requirements.txt
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
llamaapi==0.1.36
openai==1.12.0
pydub==0.25.1
python-dotenv==1.0.0
requests==2.32.3
soupsieve==2.6
urllib3==2.2.3
tiktoken==0.8.0
langfuse==2.55.0
qdrant-client==1.12.1
fastembed==0.4.2
colorama==0.4.6
neo4j==5.26.0
